Fix build after TGreatPower refactor by unifying streams and retiring helper TUs.
Move hex/zone/mission logic onto class methods, replace TStreamView with TStream,
and repair thunk call sites so MSVC500 links cleanly again.

+++ src/game/UiRuntimeContext.cpp
@@ -3,11 +3,11 @@
 // GLOBAL: IMPERIALISM 0x6a21bc
 extern "C" UiRuntimeContext* g_pUiRuntimeContext = 0;
 
-// FUNCTION: IMPERIALISM 0x00403b16
 short UiRuntimeContext::GetActiveNationId(void) {
   return activeNationIdAt2E;
 }
 
+// FUNCTION: IMPERIALISM 0x00403b16
 unsigned int __cdecl thunk_GetActiveNationId(void) {
-  return g_pUiRuntimeContext->GetActiveNationId();
+  return static_cast<unsigned int>(g_pUiRuntimeContext->GetActiveNationId());
 }
